// src/render/wave32GranaryManifest.generated.ts
export const WAVE32_GRANARY_IMAGES = {
  granary_a: {"url": "assets/wave32/granary/granary_a.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_a_boarded: {"url": "assets/wave32/granary/granary_a_boarded.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_a_empty: {"url": "assets/wave32/granary/granary_a_empty.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_a_full: {"url": "assets/wave32/granary/granary_a_full.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_a_half: {"url": "assets/wave32/granary/granary_a_half.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_a_snow: {"url": "assets/wave32/granary/granary_a_snow.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_a_weathered: {"url": "assets/wave32/granary/granary_a_weathered.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b: {"url": "assets/wave32/granary/granary_b.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b_boarded: {"url": "assets/wave32/granary/granary_b_boarded.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b_empty: {"url": "assets/wave32/granary/granary_b_empty.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b_full: {"url": "assets/wave32/granary/granary_b_full.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b_half: {"url": "assets/wave32/granary/granary_b_half.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b_snow: {"url": "assets/wave32/granary/granary_b_snow.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_b_weathered: {"url": "assets/wave32/granary/granary_b_weathered.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c: {"url": "assets/wave32/granary/granary_c.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c_boarded: {"url": "assets/wave32/granary/granary_c_boarded.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c_empty: {"url": "assets/wave32/granary/granary_c_empty.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c_full: {"url": "assets/wave32/granary/granary_c_full.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c_half: {"url": "assets/wave32/granary/granary_c_half.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c_snow: {"url": "assets/wave32/granary/granary_c_snow.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_c_weathered: {"url": "assets/wave32/granary/granary_c_weathered.png", "width": 160, "height": 144, "pivot": {"x": 80, "y": 128}},
  granary_pulley_beam: {"url": "assets/wave32/granary/granary_pulley_beam.png", "width": 64, "height": 64, "pivot": {"x": 32, "y": 64}},
  granary_pulley_rope_hook: {"url": "assets/wave32/granary/granary_pulley_rope_hook.png", "width": 32, "height": 64, "pivot": {"x": 16, "y": 64}},
  granary_sack_single: {"url": "assets/wave32/granary/granary_sack_single.png", "width": 32, "height": 32, "pivot": {"x": 16, "y": 32}},
  granary_sacks_large: {"url": "assets/wave32/granary/granary_sacks_large.png", "width": 64, "height": 64, "pivot": {"x": 32, "y": 64}},
  granary_sacks_small: {"url": "assets/wave32/granary/granary_sacks_small.png", "width": 64, "height": 48, "pivot": {"x": 32, "y": 48}},
} as const;